factor out pathnameGlob into Pattern.hs
