+++ src/Plush/Run/Expansion.hs
@@ -21,21 +21,19 @@ module Plush.Run.Expansion (
     wordExpansionActive,
     TildePart, tildeDir, tildePrefix,
     tildeSplit,
-    pathnameGlob,
     quoteRemoval,
     byPathParts,
     )
 where
 
 import Control.Applicative ((<$>))
-import Control.Monad (foldM)
-import Control.Monad.Exception (catchIOError)
 import Data.Char (isSpace)
-import Data.List (foldl', intercalate, partition, sort, tails)
+import Data.List (intercalate, partition)
 import Data.Maybe (fromMaybe)
-import System.FilePath (combine, (</>))
+import System.FilePath ((</>))
 import Text.Parsec
 
+import Plush.Run.Pattern
 import Plush.Run.Posix
 import Plush.Run.Posix.Utilities
 import Plush.Run.ShellExec
@@ -184,7 +182,6 @@ fieldSplitting ifs = expandParts $
 
 
 
-
 pathnameExpansion :: (PosixLike m) => Word -> m [Word]
 pathnameExpansion =  origIfNull $ (expandPartsM $ glob . compress)
     -- TODO: should skip pathnameExpansion of shell flag -f is set
@@ -194,65 +191,6 @@ pathnameExpansion =  origIfNull $ (expandPartsM $ glob . compress)
     glob _ = return []
 
     origIfNull f a = (\bs -> if null bs then [a] else bs) <$> f a
-
-
-data PatternPart = Literal Char | CharTest (Char -> Bool) | Star
-type Pattern = [PatternPart]
-data PatternAction = AddSlash | MatchContents Pattern
-type PathPattern = [PatternAction]
-
-pathnameGlob :: (PosixLike m) => FilePath -> String -> m [String]
-pathnameGlob base = either (\_err -> return []) match . parse patP ""
-  where
-    match :: (PosixLike m) => PathPattern -> m [String]
-    match = foldM steps [""]
-      where
-        steps fs p = concat `fmap` mapM (step p) fs
-
-        step (MatchContents p) dir = do
-            entries <- getDirectoryContents (ifNull "." $ base </> dir)
-                `catchIOError` (\_ -> return [])
-            return $ map (combine dir) $ sort $ filter (patMatch p) entries
-        step (AddSlash) path = do
-            isDir <- doesDirectoryExist (ifNull "/" $ base </> path)
-            return $ if isDir then [path ++ "/"] else []
-
-    ifNull d s = if null s then d else s
-
-    patP =  many1 (slashP <|> matchP)
-    slashP = char '/' >> return AddSlash
-    matchP = many1 partP >>= return . MatchContents
-    partP = (char '*' >> return Star)
-        <|> (char '?' >> return (CharTest (const True)))
-        <|> (char '\\' >> anyChar >>= return . Literal)
-        <|> (try $ do _ <- char '['
-                      c <- noneOf "/"
-                      first <- if c == '!'
-                               then noneOf "/"
-                               else return c
-                      rest <- many $ noneOf "]/"
-                      _ <- char ']'
-                      let f = if c == '!'
-                              then flip notElem
-                              else flip elem
-                      return $ CharTest (f $ makeCharClass (first:rest)))
-        <|> (noneOf "/" >>= return . Literal)
-
-    makeCharClass :: [Char] -> [Char]
-    makeCharClass []                   = []
-    makeCharClass (begin:'-':end:rest) = [begin .. end] ++ makeCharClass rest
-    makeCharClass (c:cs)               = c : makeCharClass cs
-
-    patMatch :: Pattern -> String -> Bool
-    patMatch pattern s = any null $ foldl' (flip concatMap) [s] matchers
-      where
-        matchers = zipWith g (True:repeat False) pattern
-        g _    (Literal l)  (c:cs)  = if l == c then [cs] else []
-        g True _            ('.':_) = []
-        g _    (CharTest p) (c:cs)  = if p c then [cs] else []
-        g _    Star            cs   = tails cs
-        g _    _               []   = []
-
 
 
 quoteRemoval :: Word -> String
